requirements.txt
aiodns==2.0.0
aiohttp==3.7.3
aiopg==1.0.0
alembic==1.4.3
async-timeout==3.0.1
attrs==20.3.0
brotlipy==0.7.0
cchardet==2.1.7
cffi==1.14.4
chardet==3.0.4
dependency-injector==4.5.1
greenlet==0.4.17
idna==2.10
iniconfig==1.1.1
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow-async==3.7.3
multidict==5.0.2
packaging==20.7
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycares==3.1.1
pycparser==2.20
pyparsing==2.4.7
pytest==6.1.2
pytest-aiohttp==0.3.0
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
typing-extensions==3.7.4.3
yarl==1.6.3
